+++ packages/abjad-convert/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 <!-- https://keepachangelog.com/en/1.0.0/ -->
 
+## [0.3.3]  2024-01-20
+### Fixed
+- letter 𐡈 in Aramaic to Arabic ط
+- add unit test to the above
+
 ## [0.3.2]  2024-01-20
 ### Fixed
 - letter 𐡇 in Aramaic to Arabic ح